+++ file-backup-s3/src/main/java/com/willmolloy/backup/s3/Main.java
@@ -5,10 +5,10 @@ import static com.willmolloy.backup.util.EnvHelper.getRequiredEnvVariable;
 import static com.willmolloy.backup.util.Preconditions.require;
 
 import com.willmolloy.backup.local.LocalStorage;
-import com.willmolloy.backup.statistics.BackupObserver;
-import com.willmolloy.backup.statistics.LoggingBackupObserver;
-import com.willmolloy.backup.statistics.discord.DiscordApi;
-import com.willmolloy.backup.statistics.discord.DiscordWebhook;
+import com.willmolloy.backup.statistics.observers.BackupObserver;
+import com.willmolloy.backup.statistics.observers.LoggingBackupObserver;
+import com.willmolloy.backup.statistics.observers.discord.DiscordApi;
+import com.willmolloy.backup.statistics.observers.discord.DiscordWebhook;
 import java.nio.file.FileSystem;
 import java.nio.file.FileSystems;
 import java.time.Duration;
